Simplify amplitude amplification: adjointable ops, folded schedule, leaner docs
Resource-estimation caching never blocked the Adj + Ctl functors. That
assumption had produced a parallel family of Adjointable* Q# operations and
an `adjointable` setting threaded through the Python mapper layer. Verified
on the simulator that a cached operation declared `is Adj + Ctl` compiles and
that its generated adjoint genuinely inverts, so the duplicates are gone and
the existing factories simply declare the wider functor set while keeping
their caching. Resource estimates are byte-identical before and after
(12 logical qubits, 244 CCZ, 1396 T-states), confirming the cache is still
live. Nothing on the Python side selects between variants any more.

Fold the round-scheduling closed forms into AmplitudeAmplification as
classmethods and delete schedule.py. Only six of its fourteen public
functions were ever used; the rest, including the BHMT exponential schedule,
are removed. The independent 2-D invariant-subspace cross-validation of the
closed forms is retained.

Optimize the QPE acceptance marker: conjugating the signal ancillas by X and
running the already-adjointable marking once, controlled on them, replaces a
flag ancilla plus a doubled comparator pass and a separate multiply-controlled
flip. Halves T-states on the Trotter path (32 -> 16). The sossa AND-ladder
reflection was prototyped and measured, but this repo's estimator does not
credit AND's measurement-based uncompute, making it a pessimization here, so
the existing reflection is kept.

Cut the user guide from 288 to 119 lines and correct every claim, name and
cross-reference to match the merged class, the `reflect_to_good_space` slot
and the removed schedule module.
