+++ python/src/qdk_chemistry/utils/qsharp/AmplitudeAmplification.qs
@@ -48,6 +48,11 @@ namespace QDKChemistry.Utils.AmplitudeAmplification {
     /// $I - 2|0\rangle\langle 0|$.  The two differ by a global phase, which is
     /// unobservable in the uncontrolled amplification loop but *is* observable
     /// when the reflection is applied as the target of a `Controlled` functor.
+    ///
+    /// Delegates to `QDKChemistry.Utils.PrepSelPrep.Reflect`, which already
+    /// special-cases the degenerate sizes (global phase for $n = 0$, a single
+    /// `Z` for $n = 1$) and, when controlled, folds the outer controls into the
+    /// same multi-controlled `Z` rather than nesting `Controlled` on top of it.
     operation ReflectAboutAllZeros(register : Qubit[]) : Unit is Adj + Ctl {
         Reflect(register);
     }
@@ -243,7 +248,7 @@ namespace QDKChemistry.Utils.AmplitudeAmplification {
     /// ## rounds
     /// Number of iterates.  Choosing this from a *lower* bound on the overlap
     /// overshoots the first maximum; see
-    /// `qdk_chemistry.algorithms.amplitude_amplification.schedule`.
+    /// `qdk_chemistry.algorithms.amplitude_amplification.AmplitudeAmplification`.
     operation ApplyAmplitudeAmplification(
         statePrep : Qubit[] => Unit is Adj,
         markingOracle : (Qubit[], Qubit) => Unit is Adj,
@@ -267,7 +272,7 @@ namespace QDKChemistry.Utils.AmplitudeAmplification {
     /// then stays inside $[1-\delta^2, 1]$ for every larger overlap, so the
     /// overshoot cliff disappears at the cost of a constant-factor increase in
     /// queries.  The phases are generated by
-    /// `qdk_chemistry.algorithms.amplitude_amplification.schedule.fixed_point_phases`,
+    /// `AmplitudeAmplification.fixed_point_phases`,
     /// which returns them in exactly the `(markPhases, statePhases)` order and
     /// reflection convention used here.
     operation ApplyFixedPointAmplitudeAmplification(
@@ -401,6 +406,14 @@ namespace QDKChemistry.Utils.AmplitudeAmplification {
     /// block encoding did not project onto the signal block, so the phase
     /// register carries no eigenvalue information for that branch.
     ///
+    /// The clean-ancilla condition is folded directly into the (already
+    /// controllable) phase-index marking: conjugating the signal ancillas by `X`
+    /// turns "all zero" into "all one", and the marking then runs *once*
+    /// controlled on them.  This avoids a separate accepted-flag qubit, the
+    /// second (uncompute) pass over the phase comparators, and the extra
+    /// multiply-controlled flip that a compute/flag/uncompute structure would
+    /// need.
+    ///
     /// # Input
     /// ## numPhaseQubits
     /// Size of the leading phase register.  The register is little-endian
@@ -420,12 +433,13 @@ namespace QDKChemistry.Utils.AmplitudeAmplification {
         let phaseRegister = register[0..numPhaseQubits - 1];
         let signalAncillas = Subarray(signalAncillaIndices, register[numPhaseQubits...]);
 
-        use accepted = Qubit();
         within {
-            ApplyAcceptedPhaseMark(phaseRegister, acceptedPhaseIndices, accepted);
             ApplyToEachCA(X, signalAncillas);
         } apply {
-            Controlled X([accepted] + signalAncillas, target);
+            Controlled ApplyAcceptedPhaseMark(
+                signalAncillas,
+                (phaseRegister, acceptedPhaseIndices, target),
+            );
         }
     }
 
@@ -464,9 +478,7 @@ namespace QDKChemistry.Utils.AmplitudeAmplification {
     /// ## controlledUnitary
     /// One adjointable callable per phase qubit, each already implementing the
     /// correct power of the walk operator.  Build these with
-    /// `QDKChemistry.Utils.PrepSelPrep.MakeAdjointableControlledPSPWalkOp`; the
-    /// resource-estimation-cached factories are not adjointable and cannot be
-    /// used here.
+    /// `QDKChemistry.Utils.PrepSelPrep.MakeControlledPSPWalkOp`.
     operation ApplyCoherentStandardQPE(
         statePrep : Qubit[] => Unit is Adj,
         controlledUnitary : ((Qubit, Qubit[]) => Unit is Adj)[],
