+++ python/src/qdk_chemistry/utils/qsharp/StatePreparation.qs
@@ -142,18 +142,8 @@ namespace QDKChemistry.Utils.StatePreparation {
 
     /// Returns a callable that applies Hadamard to each qubit in the array.
     /// # Returns
-    /// - `Qubit[] => Unit`: A callable that prepares the uniform superposition on the given qubits.
-    function MakePrepareHadamardAllOp() : Qubit[] => Unit {
-        PrepareHadamardAll(_)
-    }
-
-    /// Returns an adjointable callable that applies Hadamard to each qubit in the array.
-    ///
-    /// Coherent phase estimation and amplitude amplification reflect about the
-    /// preparation and therefore need `Adjoint`.
-    /// # Returns
-    /// - `Qubit[] => Unit is Adj + Ctl`: An adjointable uniform-superposition preparation.
-    function MakeAdjointablePrepareHadamardAllOp() : Qubit[] => Unit is Adj + Ctl {
+    /// - `Qubit[] => Unit is Adj + Ctl`: A callable that prepares the uniform superposition on the given qubits.
+    function MakePrepareHadamardAllOp() : Qubit[] => Unit is Adj + Ctl {
         PrepareHadamardAll(_)
     }
 }
